requirements.txt
numpy==1.19.2
tensorflow==2.4.1
Pillow==8.0.1
flask==1.1.2
flask_cors==3.0.10
gunicorn==20.0.4
requests==2.24.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0
click==7.1.2
python-dotenv==0.13.0
azure-cosmos==4.2.0